+++ src/Rarus/BonusServer/Transport/Formatters/Pagination.php
@@ -20,7 +20,7 @@ class Pagination
     public static function toRequestUri(?BonusServer\Transport\DTO\Pagination $pagination): string
     {
         if (null !== $pagination) {
-            return sprintf('page=%s&per_page=%s', $pagination->getPageNumber(), $pagination->getPageSize());
+            return sprintf('&page=%s&per_page=%s', $pagination->getPageNumber(), $pagination->getPageSize());
         }
 
         return '';