+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # bonus-server-php-sdk change log
 
+## 0.5.3 (31.08.2018)
+- добавлены уровни для карт [issue#61](https://github.com/rarus/bonus-server-php-sdk/issues/61)
+- для сущности `Cards` для роли `Organization` добавлен метод `getCardLevelList` [issue#61](https://github.com/rarus/bonus-server-php-sdk/issues/61) 
+- для сущности `Cards` для роли `Organization` исправлена ошибка в методе `setAccumulationAmount` [issue#61](https://github.com/rarus/bonus-server-php-sdk/issues/61) 
+- для сущности `Transactions` для роли `Organization` исправлена ошибка в методе `getSalesHistoryByCard` [issue#61](https://github.com/rarus/bonus-server-php-sdk/issues/61) 
+- для сущности `Cards` для роли `Organization` исправлена ошибка в постраничной навигации [issue#61](https://github.com/rarus/bonus-server-php-sdk/issues/61)
+- для сущности `Cards` для роли `Organization` для метода `addNewCard` добавлена возможность указать начальный баланс карты [issue#61](https://github.com/rarus/bonus-server-php-sdk/issues/61) 
+- множественные исправления в тестах
+
 ## 0.5.2 (27.08.2018)
 - исправлена ошибка c часовыми поясами и добавлены проверки в тесты `Rarus\BonusServer\Util\DateTimeParser` [issue#58](https://github.com/rarus/bonus-server-php-sdk/issues/58)